requirement.txt
Flask
flask-cors
python-dotenv
google-generativeai
db-sqlite3
pandas
openpyxl
react
react-router-dom
react-scripts
react-dom